requirements(1).txt
flask==3.0.0
gspread==6.0.2
google-auth==2.27.0
Werkzeug==3.0.1
